Fix dashboard 500 error and build issues
- Fixed duplicate lucide-react imports in dashboard page
- Downgraded Next.js from 15.3.2 to 14.2.15 for stability
- Downgraded React from 19.0.0 to 18.3.0 for compatibility
- Replaced Geist fonts with Inter and JetBrains Mono
- Recreated clean about and help pages to fix JSON input errors
- Removed duplicate/backup page files causing routing conflicts
- Added comprehensive error handling and debugging tools
- Created environment configuration guide and troubleshooting docs
- Build now successful with all pages compiling correctly
