+++ package.json
@@ -20,10 +20,10 @@
     "collab-dev": "file:",
     "lucide-react": "^0.507.0",
     "monaco-editor": "^0.52.2",
-    "next": "^15.3.1",
+    "next": "^14.2.15",
     "next-themes": "^0.4.6",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react": "^18.3.0",
+    "react-dom": "^18.3.0",
     "react-terminal": "^1.4.5",
     "socket.io-client": "^4.8.1",
     "tailwind-merge": "^3.2.0",
